Cliente.DAO modificado par usar conexão da classe de conexão

+++ src/dao/produto/Produto.DAO.pas
@@ -8,25 +8,29 @@ uses
   Model.Produto,
   DataModule,
   FireDAC.Comp.Client,
-  Data.DB;
+  Data.DB,
+  Conexao.Banco.Control;
 
 type
   TProdutoDAO = class(TInterfacedObject, iProdutoDAO)
-  constructor Create;
-  destructor Destroy; override;
+    constructor Create;
+    destructor Destroy; override;
 
-  private
-    FConexao : TDataModuleUnit;
-    FProduto : TProdutoModel;
+    private
+      FDM      : TDataModuleUnit;
+      FProduto : TProdutoModel;
+      FQuery   : TFDQuery;
 
-  public
-    procedure Salvar( aValue : TProdutoModel );
-    procedure Remover ( aValue : integer);
-    procedure Editar( aValue : TProdutoModel );
+    procedure RecuperaInstanciaQuery;
 
-    function RecuperaPorCodigo(aValue: integer; aColuna: string): Variant;
-    function VerificaSeExiste(aValue : integer ) : Boolean;
-    function RecuperaTodos : TFDMemTable ;
+    public
+      procedure Salvar( aValue : TProdutoModel );
+      procedure Remover ( aValue : integer);
+      procedure Editar( aValue : TProdutoModel );
+
+      function RecuperaPorCodigo(aValue: integer; aColuna: string): Variant;
+      function VerificaSeExiste(aValue : integer ) : Boolean;
+      function RecuperaTodos : TFDMemTable ;
   end;
 
 implementation
@@ -35,21 +39,26 @@ implementation
 
 constructor TProdutoDAO.Create;
 begin
-  Self.FConexao := TDataModuleUnit.New;
+  Self.FDM := TDataModuleUnit.New;
   Self.FProduto := TProdutoModel.Create;
 end;
 
 destructor TProdutoDAO.Destroy;
 begin
   FProduto.Free;
-  FConexao.Free;
+  FDM.Free;
 
   inherited;
 end;
 
+procedure TProdutoDAO.RecuperaInstanciaQuery;
+begin
+  FQuery := TConexaoControl.getInstancia().Conexao.CriarQuery;
+end;
+
 procedure TProdutoDAO.Editar(aValue: TProdutoModel);
 begin
-  FConexao.FDConexao.ExecSQL(
+  FQuery.ExecSQL(
     'UPDATE produto SET descricao = :descricao, preco_venda = :preco_venda WHERE codigo = :codigo',
     [aValue.Descricao, aValue.PrecoProduto, aValue.Codigo]
   );
@@ -57,7 +66,7 @@ end;
 
 function TProdutoDAO.RecuperaPorCodigo(aValue: integer; aColuna: string): Variant;
 begin
-  Result := FConexao.FDConexao.ExecSQLScalar(
+  Result := FQuery.Connection.ExecSQLScalar(
     'SELECT ' + aColuna + ' FROM produto p WHERE p.codigo = :codigo',
     [InttoStr(aValue)]
   );
@@ -65,27 +74,50 @@ end;
 
 function TProdutoDAO.RecuperaTodos: TFDMemTable;
 begin
-  FConexao.FDConexao.ExecSQL(
-    'SELECT p.codigo, p.descricao, PRINTF("R$ %.2f", preco_venda) as preco_venda FROM produto p ORDER BY codigo ',
-    TDataSet(FConexao.FDMemTable)
-  );
-
-  Result := FConexao.FDMemTable;
+  RecuperaInstanciaQuery;
+  try
+    try
+      FQuery.SQL.Clear;
+      FQuery.Open(
+        'SELECT p.codigo, p.descricao, PRINTF("R$ %.2f", preco_venda) as preco_venda FROM produto p ORDER BY codigo '
+      );
+      FDM.FDMemTable.Data := FQuery.data;
+    except
+      on E: Exception do
+        raise Exception.Create('Ops! Algo aconteceu: ' + E.Message);
+    end;
+  finally
+    Result := FDM.FDMemTable;
+  end;
 end;
+
 procedure TProdutoDAO.Remover(aValue: integer);
 begin
-  FConexao.FDConexao.ExecSQL(
-    'DELETE FROM produto WHERE codigo = :codigo',
-    [aValue]
-  );
+  try
+    if VerificaSeExiste(aValue) then
+    begin
+      FQuery.Connection.ExecSQL(
+        'DELETE FROM produto WHERE codigo = :codigo',
+        [aValue]
+      );
+    end
+  except
+    on E: Exception do
+      raise Exception.Create('Ops! Algo aconteceu: ' + E.Message);
+  end;
 end;
 
 procedure TProdutoDAO.Salvar(aValue: TProdutoModel);
 begin
-  FConexao.FDConexao.ExecSQL(
-    'INSERT INTO produto (descricao, preco_venda) VALUES (:descricao, :precovenda)',
-    [aValue.Descricao, aValue.PrecoProduto]
-  );
+  try
+    FQuery.ExecSQL(
+      'INSERT INTO produto (descricao, preco_venda) VALUES (:descricao, :precovenda)',
+      [aValue.Descricao, aValue.PrecoProduto]
+    );
+  except
+    on E: Exception do
+      raise Exception.Create('Ops! Algo aconteceu: ' + E.Message);
+  end;
 end;
 
 function TProdutoDAO.VerificaSeExiste(aValue: integer): Boolean;
@@ -93,10 +125,16 @@ var
   LRetorno : Integer;
 begin
   Result := false;
-  LRetorno := FConexao.FDConexao.ExecSQLScalar(
-    'SELECT * FROM produto o WHERE o.codigo = :codigo',
-    [InttoStr(aValue)]
-  );
+
+  try
+    LRetorno := FQuery.Connection.ExecSQLScalar(
+      'SELECT * FROM produto o WHERE o.codigo = :codigo',
+      [InttoStr(aValue)]
+    );
+  except
+    on E: Exception do
+      raise Exception.Create('Ops! Algo aconteceu: ' + E.Message);
+  end;
 
   if LRetorno > 0 then
     Result := True;
